requirements.txt
Django==3.1.2
djangorestframework==3.12.1
python-dotenv==0.10.3
requests==2.24.0
